Progressbar improvements - Python3.8 drop support
Asyncio was introduced to improve user feedback when waiting
for the response in requests using urllib.

The workflow of generating images was improved.

There is support for downloading models for inpainting

Python3.8 reached EOL.  users of older systems can continue
using previous version.

Users are encouraged to upgrade their systems, Ubuntu Mate
is one of the versions that will need to use lower versions
of the plugin.